requirements.txt
pandas
pathlib
numpy
ipython
tensorflow
django
djangorestframework
sklearn
markdown
django-filter
drf_extra_fields
pillow
django-environ
django-jazzmin
